requirements.txt
telethon
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
ffmpeg-python
